refactor(sync): split transfer and write-authority modules to satisfy Codacy code-size limits
Codacy Static Code Analysis flagged 15 new issues on PR #175: oversized
files (transfer.rs 1065 / write_authority.rs 649 non-comment lines),
oversized methods (plan 103 LOC cc18, collect_directory_stages 83 cc12,
validate_request cc11, run 67, execute_copy_file 82 cc14 10 params,
prepare_write_relative_file 67 cc13), and std::env::temp_dir usage in
test fixtures.

- split src/device/transfer.rs into a module root plus types.rs,
  planner.rs, executor.rs, and test submodules; the plan is now built
  by a PlanBuilder whose per-item methods stay under the size and
  complexity limits
- split src/local/write_authority.rs into policy.rs, staging.rs,
  target.rs, authority.rs, and tests; the conflict-policy resolution
  match is extracted from prepare_write_relative_file
- bundle executor run state into a RunContext so execute_copy_file
  takes 6 parameters instead of 10; copy_in_chunks handles the chunked
  copy with progress and cancellation
- replace std::env::temp_dir test fixtures with tempfile::tempdir()
  guards (auto-cleanup, no security-pattern match)
- PreparedWriteTarget gains a staged_path() accessor so the rollback
  and drop regressions no longer reach into private fields

No behavior change: staged file + atomic rename, per-item authority
revalidation, cooperative cancellation, reverse-order rollback, and
all error variants are preserved; every stage-ordering decision is
identical. Gates: cargo fmt --check, clippy -D warnings (debug and
release; only the 3 pre-existing newer-clippy lints in untouched
files), 1,719 locked tests pass including all 21 transfer and
write-authority regressions.
